+++ tests/serving/test_async_flow.py
@@ -15,12 +15,15 @@ import os
 import pathlib
 import shutil
 import tempfile
+import time
 import typing
 import unittest.mock
 from copy import deepcopy
+from datetime import datetime
 from types import SimpleNamespace
 from typing import Optional, Union
 
+import pandas as pd
 import pytest
 
 import mlrun
@@ -39,6 +42,13 @@ from .demo_states import *  # noqa
 class _DummyStreamRaiser:
     def push(self, data):
         raise ValueError("DummyStreamRaiser raises an error")
+
+
+def append_and_return(lst, event):
+    body = event.body
+    body["timestamp"] = datetime.now()
+    lst.append(event.body)
+    return lst
 
 
 def create_mocked_get_store_artifact(
@@ -171,6 +181,45 @@ def test_push_error():
 
     server.test(body=[])
     server.wait_for_completion()
+
+
+def test_batch():
+    function = mlrun.new_function("tests", kind="serving", project="x")
+    graph = function.set_topology("flow", engine="async")
+    step = graph
+    step = step.to("storey.Batch", "my_batching", max_events=3, flush_after_seconds=1)
+    step = step.to("storey.ToDataFrame", "my_to_df", index="my_int")
+    # to get a single result in wait_for_completion (termination result in storey)
+    step.to("storey.Reduce", initial_value=[], fn=append_and_return, full_event=True)
+    server = function.to_mock_server()
+
+    events = [{"my_int": i, "my_string": f"this is {i}"} for i in range(10)]
+
+    for event in events:
+        time.sleep(0.1)
+        server.test(body=event)
+    results = server.wait_for_completion()
+    assert len(results) == 4
+
+    prev_ts = pd.Timestamp.min
+    for i, df in enumerate(results[:4]):
+        if i < 3:
+            assert len(df) == 3, f"Batch {i} expected 3 rows, got {len(df)}"
+        if i == 3:
+            assert len(df) == 1, f"Batch {i} expected 1 rows, got {len(df)}"
+
+            # check all timestamps in the batch are the same
+            unique_ts = df["timestamp"].unique()
+            assert (
+                len(unique_ts) == 1
+            ), f"Batch {i} has multiple timestamps: {unique_ts}"
+            batch_ts = unique_ts[0]
+
+            # check timestamp order between batches
+            assert (
+                batch_ts > prev_ts
+            ), f"Batch {i} timestamp {batch_ts} not greater than previous {prev_ts}"
+            prev_ts = batch_ts
 
 
 class MyModel(Model):
